#1 - int to double

+++ src/main/java/dev/obrienlabs/mathematics/Matrix.java
@@ -2,13 +2,13 @@ package dev.obrienlabs.mathematics;
 
 public interface Matrix {
 
-	void set(int r, int c, int value);
-	int get(int r, int c);
+	void set(int r, int c, double value);
+	double get(int r, int c);
 	Matrix add(Matrix m1);
 	Matrix subtract(Matrix m1);
 	Matrix multiply(Matrix m1);
 	
-	// stats
+	// statistics
 	int rowSize();
 	int colSize();
